+++ cli/src/sdk.rs
@@ -18,14 +18,17 @@ use std::sync::Arc;
 use colored::Colorize;
 use futures_util::{SinkExt, StreamExt};
 use reqwest::StatusCode;
-use tokio::sync::mpsc::{UnboundedSender, unbounded_channel};
+use tokio::sync::mpsc::{UnboundedReceiver, UnboundedSender, unbounded_channel};
+use tokio::sync::{Mutex, MutexGuard};
 use url::Url;
 
 use self::types::*;
 use crate::config::Config;
+use crate::util::choose;
 use crate::{StrError, fail, or_fail, try_to};
 
 mod http;
+mod name;
 pub mod types;
 
 pub struct Options {
@@ -38,15 +41,18 @@ pub struct Sdk {
     options: Options,
     http: http::Client,
     user: PublicUser,
+    sessions: Mutex<Vec<PublicSession>>,
     model: ModelInfo,
 }
 
 pub struct Session {
+    id: String,
+    recv: UnboundedReceiver<Message>,
     send: UnboundedSender<tungstenite::Message>,
 }
 
 impl Sdk {
-    pub async fn new(mut options: Options) -> (Self, Vec<PublicSession>) {
+    pub async fn new(mut options: Options) -> Self {
         let mut bypass = false;
         let (http, UserInfo { user, sessions, available_models }) = loop {
             options.api_key.force(bypass);
@@ -82,9 +88,19 @@ impl Sdk {
             if options.interactive {
                 println!("Model: {}", model.model_string.green());
             }
-            return (Sdk { options, http, user, model }, sessions);
+            let sessions = Mutex::new(sessions);
+            return Sdk { options, http, user, sessions, model };
         }
         fail!("no stable model found")
+    }
+
+    pub async fn cached_sessions(&self) -> MutexGuard<'_, Vec<PublicSession>> {
+        self.sessions.lock().await
+    }
+
+    pub async fn refresh_sessions(&self) {
+        let info = try_to!("list sessions", self.http.get::<UserInfo>("/v1/user/info", None).await);
+        *self.sessions.lock().await = info.sessions;
     }
 
     async fn session_register(&self, session: &PublicSession) -> OpResult {
@@ -103,31 +119,40 @@ impl Sdk {
 }
 
 impl Session {
-    pub async fn new(sdk: Arc<Sdk>, name: String, onmessage: UnboundedSender<Message>) -> Session {
+    pub fn id(&self) -> &str {
+        &self.id
+    }
+
+    pub async fn new(sdk: Arc<Sdk>, mut name: String) -> Session {
+        if name.is_empty() {
+            name = format!("{}-{}", choose(name::ADJS), choose(name::TERMS));
+            println!("Session: {}", name.cyan());
+        }
         let session = PublicSessionBuilder::new(
             sdk.user.id.clone(),
             sdk.user.org_id.clone(),
             sdk.model.clone(),
             86400,
             name,
-            "no description".to_string(), // TODO: function paramater
+            "no description".to_string(),
         )
-        .can_log(!sdk.user.never_log) // TODO: function parameter (check can_disable_logging)
+        .can_log(!sdk.user.never_log)
         .build();
-        let id = session.id.clone();
         let result = sdk.session_register(&session).await;
         if !result.ok {
             fail!("failed to create session", &StrError(&result.status_message));
         }
         log::info!("{}", result.status_message);
-        Self::create(&sdk, id, onmessage).await
+        let id = session.id.clone();
+        sdk.cached_sessions().await.push(session);
+        Self::create(&sdk, id).await
     }
 
-    pub async fn resume(sdk: &Sdk, id: String, onmessage: UnboundedSender<Message>) -> Session {
-        Self::create(sdk, id, onmessage).await
+    pub async fn resume(sdk: &Sdk, id: String) -> Session {
+        Self::create(sdk, id).await
     }
 
-    async fn create(sdk: &Sdk, id: String, onmessage: UnboundedSender<Message>) -> Session {
+    async fn create(sdk: &Sdk, id: String) -> Session {
         let mut url = sdk.options.base_url.clone();
         url.set_scheme("wss").unwrap();
         url.set_path("/v1/stream");
@@ -137,6 +162,7 @@ impl Session {
             tokio_tungstenite::connect_async(url).await,
         );
         let (mut sink, mut stream) = stream.split();
+        let (onmessage, msg_queue) = unbounded_channel::<Message>();
         drop(tokio::spawn(async move {
             while let Some(message) = stream.next().await {
                 let message = match try_to!("receive web-socket message", message) {
@@ -157,7 +183,7 @@ impl Session {
                 try_to("send web-socket message", sink.send(message).await, None);
             }
         }));
-        Session { send }
+        Session { id, recv: msg_queue, send }
     }
 
     #[allow(dead_code)]
@@ -176,5 +202,9 @@ impl Session {
             .build();
         let message = serde_json::to_string(&message).unwrap();
         try_to!("send web-socket message", self.send.send(tungstenite::Message::text(message)));
+    }
+
+    pub async fn recv(&mut self) -> Option<Message> {
+        self.recv.recv().await
     }
 }