+++ cli/src/sdk/types.rs
@@ -240,7 +240,7 @@ api! {
     }
 }
 
-fn now() -> u64 {
+pub fn now() -> u64 {
     SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs()
 }